+++ kinto/views/records.py
@@ -59,10 +59,13 @@ class Record(resource.ShareableResource):
                            self.schema_field,
                            self.model.permissions_field)
         required_fields = [f for f in schema.get('required', []) if f not in internal_fields]
+        # jsonschema doesn't accept 'required': [] yet.
+        # See https://github.com/Julian/jsonschema/issues/337.
+        # In the meantime, strip out 'required' if no other fields are required.
         if required_fields:
             schema = {**schema, 'required': required_fields}
         else:
-            schema = {f: v for f, v in new.items() if f != 'required'}
+            schema = {f: v for f, v in schema.items() if f != 'required'}
         data = {f: v for f, v in new.items() if f not in internal_fields}
 
         # Validate or fail with 400.